refactor: reorganize registration views and update login handling
