+++ src/router/index.js
@@ -1,7 +1,6 @@
 import { createRouter, createWebHistory } from "vue-router";
 import ProductView from "@/views/ProductView.vue";
 import CartView from "@/views/CartView.vue";
-import RegisterView from "@/views/RegisterView.vue";
 import HomeView from "@/views/HomeView.vue";
 import BrandView from "@/views/BrandView.vue";
 // import ProfileView from "@/views/ProfileView.vue";
@@ -16,11 +15,10 @@ import GiftCheckoutView from "@/views/GiftCheckoutView.vue";
 import ProductAdminView from "@/views/ProductAdminView.vue";
 import { useUserStore } from "@/stores/user.js";
 import LoginView from "@/views/LoginView.vue";
-import RegisterNewView from "@/views/RegisterNewView.vue";
+import RegisterView from "@/views/RegisterView.vue";
 import ProfileTwoView from "@/views/ProfileTwoView.vue";
 import SubscriptionView from "@/views/SubscriptionView.vue";
 import MemberView from "@/views/MemberView.vue";
-
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
@@ -64,11 +62,6 @@ const router = createRouter({
       path: "/login",
       name: "Login",
       component: LoginView,
-    },
-    {
-      path: "/registerNew",
-      name: "RegisterNew",
-      component: RegisterNewView,
     },
     {
       path: "/register",
@@ -129,9 +122,9 @@ const router = createRouter({
       component: BlessingView,
     },
     {
-      path: '/order/confirm',
-      name: 'Payment',
-      component: PaymentView
+      path: "/order/confirm",
+      name: "Payment",
+      component: PaymentView,
     },
     {
       path: "/super/products/inventory",
