+++ src/views/LoginView.vue
@@ -35,9 +35,9 @@ onMounted(() => {
       client_id: clientId,
       callback: handleGoogleResponse,
       auto_select: false,
-      cancel_on_tap_outside: false
+      cancel_on_tap_outside: false,
     });
-    
+
     // 渲染 Google 登入按鈕
     window.google.accounts.id.renderButton(
       document.getElementById("google-signin-button"),
@@ -46,7 +46,7 @@ onMounted(() => {
         size: "large",
         width: "100%",
         text: "signin_with",
-        shape: "pill"
+        shape: "pill",
       }
     );
   } else {
@@ -58,21 +58,21 @@ onMounted(() => {
 const handleGoogleLogin = () => {
   // 彈出 Google 登入視窗
   window.google.accounts.id.prompt(() => {
-      window.google.accounts.id.renderButton(
-        document.getElementById("google-signin-button"),
-        {
-          theme: "outline",
-          size: "large",
-          width: "100%"
-        }
-      );
+    window.google.accounts.id.renderButton(
+      document.getElementById("google-signin-button"),
+      {
+        theme: "outline",
+        size: "large",
+        width: "100%",
+      }
+    );
   });
 };
 
 // 處理登入成功的 callback
 const handleGoogleResponse = async (res) => {
-  // console.log("Google response:", res); 
-  
+  // console.log("Google response:", res);
+
   const idToken = res.credential;
   if (!idToken) {
     alert("登入失敗 無效的 Google 憑證");
@@ -90,23 +90,39 @@ const handleGoogleResponse = async (res) => {
 </script>
 
 <template>
-  <div class="fixed inset-0 overflow-hidden bg-gradient-to-br from-[#8ecae6]/70 via-white/50 to-pink-200/70">
+  <div
+    class="fixed inset-0 overflow-hidden bg-gradient-to-br from-[#8ecae6]/70 via-white/50 to-pink-200/70"
+  >
     <!-- 註冊按鈕 -->
-    <router-link to="/registerNew"
-      class="fixed z-50 bottom-6 right-6 md:top-1/2 md:right-6 md:bottom-auto md:translate-y-[-50%]">
+    <router-link
+      to="/register"
+      class="fixed z-50 bottom-6 right-6 md:top-1/2 md:right-6 md:bottom-auto md:translate-y-[-50%]"
+    >
       <button
         class="flex items-center justify-center transition transition-transform duration-200 transform border border-white rounded-full shadow-lg w-14 h-14 bg-white/80 backdrop-blur hover:scale-125"
-        title="前往註冊">
-        <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor"
-          class="w-6 h-6 text-[#219ebc]">
-          <path stroke-linecap="round" stroke-linejoin="round" d="M13.5 4.5 21 12m0 0-7.5 7.5M21 12H3" />
+        title="前往註冊"
+      >
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          fill="none"
+          viewBox="0 0 24 24"
+          stroke-width="2"
+          stroke="currentColor"
+          class="w-6 h-6 text-[#219ebc]"
+        >
+          <path
+            stroke-linecap="round"
+            stroke-linejoin="round"
+            d="M13.5 4.5 21 12m0 0-7.5 7.5M21 12H3"
+          />
         </svg>
       </button>
     </router-link>
 
     <div class="flex items-center justify-center w-full h-full px-4">
       <div
-        class="relative w-full max-w-md p-10 bg-gradient-to-br from-white/90 to-white/70 ring-1 ring-white/40 backdrop-blur-xl hover:scale-[1.02] transition-transform duration-300 shadow-[0_10px_40px_rgba(0,0,0,0.15)]">
+        class="relative w-full max-w-md p-10 bg-gradient-to-br from-white/90 to-white/70 ring-1 ring-white/40 backdrop-blur-xl hover:scale-[1.02] transition-transform duration-300 shadow-[0_10px_40px_rgba(0,0,0,0.15)]"
+      >
         <div class="flex justify-center mb-2">
           <img src="/logo.png" alt="Kizuna Logo" class="h-10" />
         </div>
@@ -123,8 +139,12 @@ const handleGoogleResponse = async (res) => {
           <div>
             <label class="relative flex items-center text-gray-700">
               <Mail class="absolute w-5 h-5 text-gray-400 left-3" />
-              <input type="email" v-model="email" placeholder="電子郵件"
-                class="w-full pl-10 pr-4 py-3 bg-white/80 backdrop-blur rounded-full border border-gray-300 focus:border-[#219ebc] focus:ring-2 focus:ring-[#219ebc] transition-all duration-200 outline-none text-gray-800" />
+              <input
+                type="email"
+                v-model="email"
+                placeholder="電子郵件"
+                class="w-full pl-10 pr-4 py-3 bg-white/80 backdrop-blur rounded-full border border-gray-300 focus:border-[#219ebc] focus:ring-2 focus:ring-[#219ebc] transition-all duration-200 outline-none text-gray-800"
+              />
             </label>
           </div>
 
@@ -132,14 +152,20 @@ const handleGoogleResponse = async (res) => {
           <div>
             <label class="relative flex items-center text-gray-700">
               <Lock class="absolute w-5 h-5 text-gray-400 left-3" />
-              <input type="password" v-model="password" placeholder="密碼"
-                class="w-full pl-10 pr-4 py-3 bg-white/80 backdrop-blur rounded-full border border-gray-300 focus:border-[#219ebc] focus:ring-2 focus:ring-[#219ebc] transition-all duration-200 outline-none text-gray-800" />
+              <input
+                type="password"
+                v-model="password"
+                placeholder="密碼"
+                class="w-full pl-10 pr-4 py-3 bg-white/80 backdrop-blur rounded-full border border-gray-300 focus:border-[#219ebc] focus:ring-2 focus:ring-[#219ebc] transition-all duration-200 outline-none text-gray-800"
+              />
             </label>
           </div>
 
           <!-- 登入按鈕 -->
-          <button type="submit"
-            class="w-full py-3 text-white font-semibold rounded-full bg-gradient-to-r from-primary to-pink-300 hover:from-[#7bb8d9] hover:to-pink-400 shadow-lg hover:shadow-xl transition-all duration-200">
+          <button
+            type="submit"
+            class="w-full py-3 text-white font-semibold rounded-full bg-gradient-to-r from-primary to-pink-300 hover:from-[#7bb8d9] hover:to-pink-400 shadow-lg hover:shadow-xl transition-all duration-200"
+          >
             登入
           </button>
         </form>
@@ -153,11 +179,18 @@ const handleGoogleResponse = async (res) => {
 
         <!-- Google 登入 - 使用官方按鈕 -->
         <div id="google-signin-button" class="w-full"></div>
-        
+
         <!-- 備用的自訂按鈕 -->
-        <button @click="handleGoogleLogin" v-if="false"
-          class="flex items-center justify-center w-full gap-3 py-3 text-sm font-semibold rounded-full border border-gray-300 bg-white/80 backdrop-blur hover:bg-white hover:border-[#219ebc] transition-all duration-200">
-          <img src="https://www.svgrepo.com/show/475656/google-color.svg" alt="Google" class="w-5 h-5" />
+        <button
+          @click="handleGoogleLogin"
+          v-if="false"
+          class="flex items-center justify-center w-full gap-3 py-3 text-sm font-semibold rounded-full border border-gray-300 bg-white/80 backdrop-blur hover:bg-white hover:border-[#219ebc] transition-all duration-200"
+        >
+          <img
+            src="https://www.svgrepo.com/show/475656/google-color.svg"
+            alt="Google"
+            class="w-5 h-5"
+          />
           <span class="text-gray-800">使用 Google 登入</span>
         </button>
       </div>
